+++ src/main/java/baseball/result/GameResult.java
@@ -1,0 +1,31 @@
+package baseball.result;
+
+public class GameResult {
+    private int ballCount;
+    private int strikeCount;
+    private String resultMessage;
+
+    public int getBallCount() {
+        return ballCount;
+    }
+
+    public void setBallCount(int ballCount) {
+        this.ballCount = ballCount;
+    }
+
+    public int getStrikeCount() {
+        return strikeCount;
+    }
+
+    public void setStrikeCount(int strikeCount) {
+        this.strikeCount = strikeCount;
+    }
+
+    public String getResultMessage() {
+        return resultMessage;
+    }
+
+    public void setResultMessage(String resultMessage) {
+        this.resultMessage = resultMessage;
+    }
+}